requirements.txt
asgiref==3.6.0
axios==0.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Django==4.1.7
django-cors-headers==3.14.0
django-filter==23.1
django-rest-framework==0.1.0
django-tables2==2.5.3
djangorestframework==3.14.0
et-xmlfile==1.1.0
idna==3.4
lxml==4.9.2
markdown-it-py==2.2.0
mdurl==0.1.2
mysqlclient==2.1.1
numpy==1.24.3
openpyxl==3.1.2
pandas==2.0.0
Pygments==2.15.1
python-dateutil==2.8.2
pytz==2023.3
requests==2.28.2
rich==13.3.4
six==1.16.0
sqlparse==0.4.3
tzdata==2023.3
urllib3==1.26.15
xlrd==2.0.1
xlwt==1.3.0
Django==4.1.7
django_filter==23.1
django_phonenumber_field==7.1.0
django_tables2==2.5.3
djangorestframework==3.14.0
pandas==2.0.0
python_dateutil==2.8.2
pytz==2023.3
xlrd==2.0.1
xlwt==1.3.0
